fixup! just sync test-case file into LLVM; don't move tests
